Fixes session end logging issues with an extraneous temporary script.
Tmux does not provide a good pre-kill hook/event. The session-closed hook is not
dependable since it fires after session exit at which point, options set on the
session are no longer available.

While a circuitous option, the closest possible approach right now seems to be
a temp shell script with the onexit workflow and a TRAP on exit while creating the session.